+++ src/convtools/contrib/models/validators/__init__.py
@@ -1,3 +1,3 @@
 """Exposes validators - objects which only check the data and populate error
 objects"""
-from .validators import Custom, Gt, Gte, Lt, Lte, Regex, Type
+from .validators import Custom, Decimal, Gt, Gte, Lt, Lte, Regex, Type